+++ src/util.ts
@@ -42,13 +42,6 @@ export const DEFAULT_DEBUG_ARTIFACT_NAME = "debug-artifacts";
  */
 export const DEFAULT_DEBUG_DATABASE_NAME = "db";
 
-/**
- * Environment variable that is set to "true" when the CodeQL Action has invoked
- * the Go autobuilder.
- */
-export const DID_AUTOBUILD_GO_ENV_VAR_NAME =
-  "CODEQL_ACTION_DID_AUTOBUILD_GOLANG";
-
 export interface SarifFile {
   version?: string | null;
   runs: SarifRun[];
@@ -312,10 +305,12 @@ export enum GitHubVariant {
   DOTCOM,
   GHES,
   GHAE,
+  GHE_DOTCOM,
 }
 export type GitHubVersion =
   | { type: GitHubVariant.DOTCOM }
   | { type: GitHubVariant.GHAE }
+  | { type: GitHubVariant.GHE_DOTCOM }
   | { type: GitHubVariant.GHES; version: string };
 
 export async function getGitHubVersion(
@@ -339,6 +334,10 @@ export async function getGitHubVersion(
 
   if (response.headers[GITHUB_ENTERPRISE_VERSION_HEADER] === "GitHub AE") {
     return { type: GitHubVariant.GHAE };
+  }
+
+  if (response.headers[GITHUB_ENTERPRISE_VERSION_HEADER] === "ghe.com") {
+    return { type: GitHubVariant.GHE_DOTCOM };
   }
 
   const version = response.headers[GITHUB_ENTERPRISE_VERSION_HEADER] as string;