+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # CodeQL Action Changelog
 
+## 2.2.8 - 22 Mar 2023
+
+- Update default CodeQL bundle version to 2.12.5. [#1585](https://github.com/github/codeql-action/pull/1585)
+
 ## 2.2.7 - 15 Mar 2023
 
 No user facing changes.
 
 ## 2.2.6 - 10 Mar 2023
 
-- Update default CodeQL bundle version to 2.12.4.
+- Update default CodeQL bundle version to 2.12.4. [#1561](https://github.com/github/codeql-action/pull/1561)
 
 ## 2.2.5 - 24 Feb 2023